Resolves several compile warnings when comparing different int signedness (2)
- advances #713 and eliminates all warnings concerning this operation:
- in R_dsgraph_structure::load() / r__dsgraph_build.cpp
- var 'vcnt' converted to u32 in R_Backend_DBG.cpp
- several for loops in light.cpp
- several for loops in in DetailManager_CACHE.cpp
- var 'filled_blocks' converted from int to i32, in SoundRender_Emitter.cpp
- in CSoundRender_Core::create() / SoundRender_Core.cpp,
explicit convertion of var 'sg_SourceType' to int
- for loop in CollectorPacked::CollectorPacked() / xrCDB_Collector.cpp
- several for loops in Frustum.cpp

+++ src/Layers/xrRender/DetailManager_CACHE.cpp
@@ -37,9 +37,9 @@ void CDetailManager::cache_Initialize()
 CDetailManager::Slot* CDetailManager::cache_Query(int r_x, int r_z)
 {
     int gx = w2cg_X(r_x + cache_cx);
-    VERIFY(gx >= 0 && gx < dm_cache_line);
+    VERIFY(gx >= 0 && gx < static_cast<int>(dm_cache_line));
     int gz = w2cg_Z(r_z + cache_cz);
-    VERIFY(gz >= 0 && gz < dm_cache_line);
+    VERIFY(gz >= 0 && gz < static_cast<int>(dm_cache_line));
     return cache[gz][gx];
 }
